+++ pkg/nazahttp/http.go
@@ -10,12 +10,30 @@ package nazahttp
 
 import (
 	"io"
+	"io/ioutil"
 	"net/http"
 	"os"
 	"time"
 )
 
-// 获取 http 文件保存至本地
+// 获取http文件
+// @return <error> 成功返回nil
+func GetHttpFile(url string, timeoutMSec int) ([]byte, error) {
+	var c http.Client
+	if timeoutMSec > 0 {
+		c.Timeout = time.Duration(timeoutMSec) * time.Millisecond
+	}
+	resp, err := c.Get(url)
+	if err != nil {
+		return nil, err
+	}
+	defer resp.Body.Close()
+
+	content, err := ioutil.ReadAll(resp.Body)
+	return content, err
+}
+
+// 获取http文件保存至本地
 func DownloadHttpFile(url string, saveTo string, timeoutMSec int) (int64, error) {
 	var c http.Client
 	if timeoutMSec > 0 {